refactor(editor): extract the shared BlockNote schema into @memry/editor-schema
One factory both processes build from, so neither can carry a node type the
other lacks. wikiLink and linkMention move whole (vanilla DOM, no renderer
imports); hashTag and dateMention move config + parse + toExternalHTML only --
their renders reach for the tag palette, HugeIcon loader and clock/week-start
settings, so each process supplies presentation.

Renderer behaviour is unchanged: 609 test files / 6965 tests green, and the
lockfile diff is only the new package's own entry (same @blocknote/core patch
hash, so still one schema identity). Main is switched over in the next commit.
