+++ packages/editor-schema/src/index.ts
@@ -1,0 +1,2 @@
+export * from './inline'
+export * from './schema'
