+++ packages/editor-schema/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "@memry/typescript-config/node.json",
+  "compilerOptions": {
+    "lib": ["ES2022", "DOM"]
+  },
+  "include": ["src/**/*"],
+  "exclude": ["**/*.test.ts", "**/*.test.tsx", "**/*.spec.ts", "**/*.spec.tsx"]
+}
